[IMP] Adjust dependency
