+++ account_invoice_line_quant/models/__init__.py
@@ -1,3 +1,2 @@
 from . import account_invoice
 from . import account_invoice_line
-from . import sale_order_line
