+++ account_invoice_line_quant/models/account_invoice_line.py
@@ -1,7 +1,7 @@
 # Copyright 2019 Quartile Limited
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
 
-from odoo import api, fields, models
+from odoo import fields, models
 
 
 class AccountInvoiceLine(models.Model):
